Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

Can we just use a different gerrit url instead of a github URL here? github is less reliable and has more of a chance of getting a 404 when we're running stack.sh as part of a an automated job, which would lead to false test failures.

Patch-set: 2
Label: Code-Review=-1