Enhanced identification of HOST_IP and eth device
Enhance the algorithm to automatically identify the HOST_IP
and NIC to be used for networking, simplifying things so that,
for most uses, defaults will `just work'.

First use the first NIC that has a valid IP address.  If there
is no such device then use the IP from the `my_ip' field of the
`nova.conf' file.  If that fails then inform the user that the
script can't find a NIC and bail.

Change-Id: I8798395fb73b379508ba74ede83908de766572b0
Signed-off-by: Don Dugger <donald.d.dugger@intel.com>