Update patch set 1
Patch Set 1:

We already assume that stack.sh is present, or rather that properly configured stackrc/localrc is present.  I have a config path for when UPLOAD_LEGACY_TTY is set that uses the tty image. The tests themselves make assumptions that are wrong with both approaches.  I can go either way with it.

Patch-set: 1
Label: Code-Review=0