Update patch set 4
Patch Set 4:

This will leave the git tree in detached head state, which is probably fine -- depends on whether the next thing you expect to do with the tree is throw it away, or commit a change.  

However, if the use case for this is to test out a change, and possibly update it, then it may be worth having devstack simulate exactly what git-review would do in this case, including checking out onto a specially named branch, such that a subsequent git-review is run, it is uploaded with the same topic.

One way to implement this would be to use git-review.  Perhaps the following modification to this change:

If 'git-review' is in the path, use "git review -d CHANGEID" instead of the checkout/fetch, if not, leave the code as-is.

I'm not sure whether this is a use case you're targeting, and whether you would expect SSH keys, etc, to be available for git-review in the target environment.

Patch-set: 4