Update patch set 2
Patch Set 2:

I'm not sure how to get the default value for volume_name_template before nova starts running if it isn't in the flags file.  The alternative of always needing to set it in the flags file based on VOLUME_NAME_TEMPLATE occurred to me and is probably the right solution but it felt like overkill for something that will rarely be changed in practice.

Patch-set: 2