Update patch set 1
Patch Set 1: No score

The reason for not including pip-requires from the projects is that they sometimes aren't sane for development.

Each project takes a difference stance on what it means to be in tools/pip-requires.  Sometimes it includes things that are useful for dev - like nose - sometimes there are pips that are optional...  Sometimes projects have (or had) a tools/pip-requires-dev for dev and the normal pip-requires is for running it

Patch-set: 1
Reviewer: Gerrit User 112 <112@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=0