Update patch set 6
Patch Set 6:

No log file is now the default unless LOGFILE is set.  A '.xxxxxxxx' is appended to LOGFILE where xxxxxxxx is a representation of the date.  Also cleanup occurs unless LOGFILE is set, the value of LOGFILE is appended with '.*' to do the cleanup in the directory specified by the user only.

I removed the '-' usage as it was only used to override the default log file name in the previous behaviour.

Patch-set: 6
Label: Code-Review=0