Update patch set 7
Patch Set 7:

Added a link from user's LOGFILE to the current logfile+timestamp

Scott, I wanted to keep LOGFILE as-is so it can be parsed for the find to clean up old files so I added yet another var TIMESTAMP_FORMAT so the date format can be overridden.  It is still appended to LOGFILE.  I hope this is close enough.

Patch-set: 7