Update patch set 2
Patch Set 2:

Hi Dean.  That's a good point.  I'm wondering if it wouldn't be cleaner to move all of that code under the "enable quantum in dashboard, if requested"  if-block.  

My guess is that horizon is fine with the openstack-dashboard/quantum directory being missing if Quantum is not enabled, but we're probably setting up a bum symlink if Quantum is not installed.  

Now that Quantum has a separate client lib, we're going to work with the Horizon folks to make this cleaner.

Patch-set: 2