Update patch set 2
Patch Set 2:

Well it doesn't need it now but presumably you would only need the quantum client on the host that is running horizon, but not on the host running the nova-network service. I'll leave it as is in the other merge for now, since I don't use horizon or multi-node someone else can make that call ;).

Patch-set: 2