Update patch set 1
Patch Set 1:

My hope is that the stuff in 'functions' is not mysterious enough to disrupt the flow of reading stack.sh. I share the concern of having to look in too many places when reading scripts (re. RedHat's init.d script spaghetti), but think that a little normalization and housekeeping balances that here.

Patch-set: 1
Label: Code-Review=0
Label: Workflow=0