Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

IMHO: This is a fairly big change to support the lucid distribution. 

This is nice to have another distro supported but I would say that devstack is mostly designed for bleeding edge distributions and should be targeted for the only latest.

Another argument would be that if we add LTS support this could maybe encourage people to run on production which defeat the point of devstack.

Adding this patch as well would add more complexities to the script which by design I would say the script should be keep as lean and readable as possible.

Patch-set: 1
Reviewer: Gerrit User 866 <866@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1