Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

This only looks at the user's default shell, not the one currently running.  screen uses $SHELL to decide what shell to use in a new window so it seems to be that is what is important here.

Patch-set: 1
Label: Code-Review=-1
Label: Workflow=0