Update patch set 1
Patch Set 1:

This is a merge commit that will completely replace the contents
of the "master" branch with the contents of the "redux" branch,
which is designed to work with "keystone light". Run "git review
-d 4172" to see what that will look like.

When you're ready for this to be merged and become master,
approve the change in the normal way.  We know that this branch
works with the redux branch of keystone, but since we can only
merge one change at a time, this will fail the devstack test.
Once the merge proposals in both projects are approved, we'll
bypass the devstack test and merge them.

Patch-set: 1
Reviewer: Gerrit User 1 <1@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=0
Label: Workflow=0