Update patch set 2
Patch Set 2:

talking w/ Gabe and Tres - thought we'd need it to let user's change their passwords, but Gabe's thinking not - and to put in a keystone patch later to let the "is_admin" check for user updates.

The question was "do we need a service account for Horizon". The gist right now from #openstack-horizon is that we'll leave in the account in case we need it, but we don't want to stash the credentials.

Patch-set: 2
Label: Code-Review=0