Update patch set 4
Patch Set 4: I would prefer that you didn't submit this

(7 inline comments)

Overall, you need to follow the structure of the existing exercise scripts; admittedly we changed that a bit in the last few days to make debugging failures a bit easier.  See swift.sh for an example of the banners and error detection we added.

Also, the exercise scripts should only use the client binaries to interact with OpenStack as demonstrating their use is one of their principal functions.  If certain configuration needs to be present for the test to operate, it should be in stack.sh, or called from stack.sh (see keystone_data.sh for an example of this).

And lastly, the exercise script should clean up after itself if successful.

Patch-set: 4
Label: Code-Review=-1
Label: Workflow=0