Update patch set 2
Patch Set 2:

And as I see it, the only reason for this patch is to find stackrc/localrc to get  HOST_IP/SERVICE_HOST.  What if stack.sh created openrc from a template and stuffed in either the HOST_IP or the path to stackrc?

I like having a single place for setting up user auth, I don't think we need an exercise-specific rc script for that.

Patch-set: 2
Label: Code-Review=0