Add support for local.sh and unstack.sh
Run local.sh at the end of stack.sh if it exists and is executable.
This allows the user to automatically perform local actions on every
re-stack, such as creating custom flavors or specific tenants/users.
Like localrc, this file is not distributed with DevStack so user
modifications will be undisturbed.

unstack.sh is a simple clean-up script to kill known running
OpenStack processes.

is_service_enabled() moves to functions

Change-Id: I41b33817c4436e644c336e4e0673144ac0844c26