Update patch set 6
Patch Set 6:

You don't need to touch the errexit (-e) setting at all.  This is the first (I think) case in the exercises of wanting to die on a zero exit code.  So use the logical and instead:

 <successful-command> && die "command succeeded"

And as used in the other exercises, to die on a non-zero exit code:

 <failed-command> || die "command failed"

And of course, to keep executing the script on a non-zero exit code:

 <failed-command> || true

Patch-set: 6
Label: Code-Review=0
Label: Workflow=0