Update patch set 1
Patch Set 1:

Scott, that's what I had in mind.  The stackrc bits probably need to be in stack.sh as the UPLOAD_xxx_TTY vars are not set yet.  But if the contents of the 'for image_url' loop become a function then Dave's change becomes a call to that function and IMAGE_URLs is not an issue.  Same for the ansolabs tty image.

Patch-set: 1