Update patch set 8
Patch Set 8: (1 inline comment)

I think this was mentioned before, but generally exercises use rest apis, and should be runnable on a machine outside the cluster.  So we try not to use [x]-manage commands, which require privileged conf files and network access to dbs etc.  Similar rationale applies to direct mysql calls.

As for the existing nova-manage calls, some of this may be possible to do through nova's api:

https://github.com/openstack/nova/blob/master/nova/api/openstack/compute/contrib/networks.py#L94

Also, is it possible that you could use networks.show to replace your raw mysql query?

Patch-set: 8
Label: Workflow=0