Update patch set 8
Patch Set 8:

It looks like there is a some api/cli stuff to figure out both for nova and quantum here, but those are larger Folsom issues.  My recommendation is to wait till essex is cut, and then merge this as-is.  Even if it is not perfect, it is useful in that it brings to light the user-facing issues (for which we should log bugs/blueprints) and will make it easier to collaborate to fix those issues.

Patch-set: 8