Depricate 'm-svc', use 'melange'
There is only one Melange service, no need for the prefixed service name for
grouping that is used by Glance, Nova and Quantum.  Just use 'melange' in
ENABLED_SERVICES.

There is a DEPRICATED message displayed at the end of stack.sh as a reminder
for users to update their configurations.

Fixes bug 968537

Change-Id: Id93546ef73db875854aff55a09c212badb5ce7c9