Update patch set 2
Patch Set 2:

Note: I'm not a huge fan of this fix, as it removes the files from the directory where the source is kept, but I thought I would try to find a fix to the issue that did not involve reverting the original patch (https://github.com/openstack-dev/devstack/commit/17ff9763da677f3db0339a395c558cdb881d5d56).  

I'd be OK with just revert of the original patch as well, which would move this back to using the config files in the source directory for the time being.  I'm not sure what the original motivation was for changing it, so I can't comment on the downsides of completely reverting.

Patch-set: 2
Reviewer: Gerrit User 447 <447@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=0