Update patch set 3
Patch Set 3:

Hi Dean,

Let me know if you were able to execute the tests successfully  using devstack(without my changes) with "quantum enabled". I am not able to, so the checkin is probably not related to the error.

Would you confirm? And If that is the case this checkin could be approved right? because the manual creation and deletion of the vms in specific networks within quantum is working as expected with my changes.

Please let me know

Thanks
Shweta

Patch-set: 3