Update patch set 8
Patch Set 8:

Talking with Vish, our target is to have Quantum be the default networking mode in Folsom, but not if we have concerns around stability.  We'll probably make a final call around the start of F-4.  

There's going to be a fair amount of churn in Quantum and the Nova + Quantum integration code in the early F milestones, and I don't want that to cause problems for other people.

Patch-set: 8
Reviewer: Gerrit User 447 <447@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=0