Update patch set 8
Patch Set 8: I would prefer that you didn't submit this; No score

We recently added a facility to skip exercise scripts if a required service is not enabled.  See exercises/swift.sh around line 40 for an example.  This will take this script out of the gating process when quantum is not enabled.

And will automatically add it back when we default to quantum enabled.

Patch-set: 8
Label: Code-Review=-1
Label: Workflow=0