Update patch set 4
Patch Set 4: Do not submit

(2 inline comments)

errexit does not cause the script to exit when a command returns non-zero exit code when the command is in a sub-shell, such as in backticks.  You need to do something like this to test for command failure:

    MEMBER_ID=`keystone role-list | grep Member | cut -d '|' -f2 | tr -d ' '`
    die_if_not_set MEMBER_ID "failure getting member id"

This fails immediately because keystone requires admin privs for almost everything and the default in openrc is to set up the demo account.  The die_if_not_set function is in 'functions':

    source ./functions

To get admin privs:

    source ./openrc admin admin

Patch-set: 4
Reviewer: Gerrit User 970 <970@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-2
Label: Workflow=0