Update patch set 3
Patch Set 3: I would prefer that you didn't submit this

sweet, thanks. 

This has thumbs up from me apart the image type issue: if an ami is not available the script exits with 0 and says that it has finished configured Tempest. Technically it has, but the tempest.conf is broken ;)

Patch-set: 3
Label: Code-Review=-1