Allow wget to handle Non-English output
The stack.sh has a loop to wait for keystone to start. But the output of wget tool has been globalized which means it won't return the English word 'refused' on Non-English environment even the keystone is not up. So the script will assume the keystone has been started and continue. The command of keystone tenant-create always failed immediately after skipping this loop since it require keystone to be started to authenticate. That's why you can see authentication error in the log because the tenant information was not correctly set up.

Fix bug 978739

Change-Id: Ia4ebe6ad0f9b5a886de48a4bea7c2aebff420dad