Update patch set 2
Patch Set 2:

Hi Dan,

I just tried to setup the linuxbridge plugin using this script.
I see an error issued when a vm is created on the default network. It looks like it is creating the bridge interface for the default network before the first vm is created. So the vm creation on the default network fails

I checked with Sumit and he said he had seen a similar error while reviewing this patch.
https://review.openstack.org/#/c/7169/
So I am not sure if it is linked to your changes

Vms created on private networks seem to go through ok.

Patch-set: 2