Update patch set 2
Patch Set 2:

Dan, part of the problem here has been multiple patches on the same bit of code.  The other is that I can't test all of the permutations so I rely on folk closer to the project to do some of that and give a positive review.

And, guess what...it needs another rebase.  I promise, no more quantum changes to stack.sh until this one gets in!  ;)

Patch-set: 2
Label: Code-Review=0
Label: Workflow=0