Update patch set 2
Patch Set 2: Looks good to me, but someone else must approve

Thanks Dan for doing this. I have tested this with the Linux Bridge plugin on a single host, and things seem to be working fine.

There is a patch I have submitted for review which related to an issue in the LB interface driver, and which I had to use here to test this.

I see an issue with the VMs on the default public network not being able to reach the meta-data server. I have observed this even in the trunk and I don't believe it has anything to do with this patch.

Patch-set: 2
Label: Code-Review=+1