Update patch set 3
Patch Set 3:

btw, thanks for the comments Dean.

One thing that will really help is getting Quantum into the commit gating, so we can be sure that changes to the devstack script do not break Quantum prior to those changes being merged.  We're definitely working on this, and hopefully it will make your life easier.  Devstack is a huge win for the Quantum team, so keeping it in good shape is a priority for us.  

Let me know how I can help out, and feel free to add me to Quantum-related devstack reviews.

Patch-set: 3