Update patch set 1
Patch Set 1:

I think it would be good to get essex stable working with the new style of XenServer install.

I want to get some qa running with XenServer and essex stable, and this change would be a big help.

Really just pushing this to gain an initial yes/no for the idea.

Patch-set: 1
Label: Code-Review=0