Update patch set 1
Patch Set 1:

Renuka spent some time trying to track down what that was. I think it is because those args are copied by the installer into grub, then supplied again on the following boots, but haven't confirmed that or tried to fix it.

I can add that change in. In my system I want it to work cross machine, so I just specify the bridge name, so I didn't see this - oops.

Patch-set: 1