Update patch set 1
Patch Set 1: Looks good to me (core reviewer)

looks good.  We really should make this write out HOST_IP to the config when it autodetects it so that it works on subsequent runs. (On subsequent runs it won't find it because it has been moved to br100)

Patch-set: 1
Label: Code-Review=+2
Label: Workflow=0