Update patch set 1
Patch Set 1:

For the record, I think we should set DISTRO in GetOSVersion, which would remove the need for checking if it's set in get_packages(). But this can be done in a later change.

Also, get_packages() now deals with cinder which is something that should probably be kept for https://review.openstack.org/#/c/7042/

(I'd move DEST to stackrc in a later commit, but that's just me ;-))

Patch-set: 1
Reviewer: Gerrit User 4460 <4460@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=0