Add ERROR_ON_CLONE.
This lets the user assert that stack.sh should never need to clone
any git repositories.  If set to True, and devstack does need to
clone a git repo, stack.sh will exit with an error.

This is useful in testing environments to make sure that the correct
code is being tested instead of silently falling back on cloning
upstream.

Change-Id: Ic0312ab4df492c5cf2e04c08aa7669a81736daa6