Update patch set 2
Patch Set 2: I would prefer that you didn't merge this

Make the additional three changes in 8861 here instead since this patchset was not merged.  Make sure the Change-Id in the comment block is the same as it is now.

Remove the change to stack.sh as 8018 will handle that.

Also, note that 8018 has an inline comment suggesting that the 'head' in the awk pipeline be replaced with an exit in the awk code.

Patch-set: 2
Reviewer: Gerrit User 970 <970@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1
Label: Workflow=0