Update patch set 8
Patch Set 8: (2 inline comments)

Hi,
Sorry for being a pain. I am reviewing your v2 work for the ovs and as a result you can remove the -a for the agents. I am doing the same support for the linuxbridge.
Keep up the great work.
Thanks
Gary

Patch-set: 8
Label: Code-Review=0