Create change
Patch-set: 1
Change-id: Icc2bd9551e235173683e3677880c2641faee62a9
Subject: The stack.sh has a loop to wait for keystone to start. But the output of wget tool has been globalized which means it won't return the English word 'refused' on Non-English environment even the keystone is not up. So the script will assume the keystone has been started and continue. The command of keystone tenant-create always failed immediately after skipping this loop since it require keystone to be started to authenticate. That's why you can see authentication error in the log because the tenant information was not correctly set up.
Branch: refs/heads/master
Commit: 04c180fe7d07b7974105965da90d306f8d08b1ad