Update patch set 10
Patch Set 10:

In this commit I changed the mv's to cp's which dealt with moving the quantum configuration around. The reason for this is because git always thinks I deleted these files when I go to make a commit. They also don't need to be mv's because all the quantum commands reference the correct configuration file we want to use.

Patch-set: 10
Label: Code-Review=0