Search multiple interfaces for host ip
When running devstack, nova moves the host ip from eth0 onto the
bridge. This causes devstack to fail on the second run unless you
explicitly set HOST_IP in localrc.

This patch searches for an ip on br100 before falling back to
other interfaces so that we can run devstack multiple times
without manually setting HOST_IP in localrc.

It also will filter out ips from the potential list that are part
of the fixed range and floating range if the netaddr library is
installed. This allows us to find the proper ip even if we have
accidentally left a floating ip or fixed ip on the bridge.

Change-Id: I13288e53ee2786c5ae0edb3f9ab457be8303f1f6