Update patch set 2
Patch Set 2: Looks good to me, but someone else must approve

This is REALLY an issue under non-western language system. If the physical machine is fast enough to launch the keystone service in a short time, there won't be this issue. However, if the physical machine does not perform well(e.g. The keystone service is not ready, but the keystone_data.sh is being executed), this issue will be raised. This patch can solve it. I hope someone can give it +2.

Patch-set: 2
Label: Code-Review=+1