Update patch set 1
Patch Set 1:

You could have ex2_create_image look at VOLUME_BACKING_FILE_SIZE and decide if it has enough room to run and exit with code 55 if it doesn't.  That is the magic code for an exercise script to skip itself and continue exercise.sh execution.  That would let it run by default when it can at least.

Patch-set: 1