Improvements to quantum.sh exercise script
Many small fixes to the quantum.sh script to make it work reliably
and with the latest versions of various clients.

Note: these changes are still using Quantum v1, as v2 does not yet
fully support this test.

Updates include:
- adding a delay between booting VMs (usely on double-virt systems)
- using nova-manage, not mysql access to get network-ids.
- using 'source openrc' to set correct credentials for each operation
- add "fast exit" checks to see when servers become active, rather than
a static pause.
- changed IP prefix values to be less likely to collide with other
networks used for testing.
- poke security group holes to allow pings.
- add pause between killing VMs and deleting networks to avoid race
where network tear-down fails b/c there are still ports on the network.
- change behavior of script with no arguments to run all tests (rather
than requiring -a)
- add flags to selectively run tear-down portions of the script.  Useful
for incremental testing/debugging.
- change --flagfile to --config-file for nova-manage
- remove admin role from test tenants demo1 and demo2 (this was braking
an attempt to add the security group, as novaclient assumes security
group names are unique per query, which is not true for admin users).

Change-Id: If04ca7854fedf485db8de31c5b18264056dad2af