Update patch set 3
Patch Set 3:

I notice a lot of timeout values are used by just sleeping for that amount of time.  I believe devstack and exercises usually set larger timeouts, and then perform repeated actions to determine when a condition is met.  In other words, instead of doing:

  sleep 15
  # assume server is destroyed

It does:

  until 120 second timeout is reached:
    if server is destroyed: break
    sleep 1

Which makes it more resilient in case the tests are running on a slow machine.  Since we run the devstack-gate on VMs that can be very slow at times, I worry about whether just sleeping and then assuming an action is complete is resilient enough -- perhaps the exercises should adopt the busy-wait approach with longer timeouts.

(Also, if you sleep 15 seconds but you only needed to wait 2, it's just wasting time.)

Patch-set: 3
Label: Verified=0
Label: Code-Review=0
Label: Workflow=0