Update patch set 2
Patch Set 2: (3 inline comments)

Salvatore, Dan, thanks for the advice.

Regarding the pipeline configuration, I don't think so it would be too difficult to introduce the configuration option for the auth pipeline. It will require to modify quantum code itself. Is my understanding right?

We need to decide the direction whether to use the current adhoc approach at this moment temporarily or to modify ption for auth strategy to quantum. If the latter is adopted, i am happy to work on this.

I will appreciate advices for the direction.

Patch-set: 2
Label: Code-Review=0