Update patch set 4
Patch Set 4:

I don't think any of the ceilometer developers are using F17. Since we're not enabling the service by default here, could we go ahead with the patch and circle back if we can get someone to help us test and debug? Or should we disable it, and then have whoever is doing that debugging re-enable it?

Patch-set: 4