Update patch set 8
Patch Set 8: (1 inline comment)

It is good, because
Precise + DHCP will work !

Precise + FixedIP will leave you with no DNS servers inside the guest.
It might be an idea to detect that case, and do an exit 1, so that the user is notified. (Exotic configuration, if someone needs it, fix it).

Patch-set: 8
Label: Code-Review=0