Update patch set 6
Patch Set 6:

Actually I think you're ready for multi-node.  screen_it already does individual checks and you've added the catch-all for the project name.  You should be able to put the individual names in ENABLED_SERVICES to get less than all of them.  --untested by me--

On another note, if F17 is not supported can there be a sanity check to at least warn that it is unsupported and maybe to not start the services?  Since this isn't core I'd be OK with letting stack.sh continue if ceilomenter doesn't start in this case.

Patch-set: 6