Add support to optionally launch the heat service.
This draft review is to give the heat developers something to comment on,
and to give the devstack developers a preview of heat support.

This allows the heat service to be started as a devstack service.

Heat is disabled by default, and can be enabled with this in your localrc:
ENABLED_SERVICES+=,heat,h-api,h-eng,h-meta

There is now a repo of heat-enabled images here:
https://github.com/heat-api/prebuilt-jeos-images/downloads

These can be added to the IMAGE_URLS in your localrc.

After devstack is launched, a template can be invoked with:
nova keypair-add --pub_key $HOME/.ssh/id_rsa.pub heat_key
heat -d create wordpress \
--template-file=../heat/templates/WordPress_Single_Instance.template \
--parameters="InstanceType=m1.tiny;DBUsername=wpuser;DBPassword=wppassword;\
KeyName=heat_key;LinuxDistribution=F16"

Change-Id: I07591295eb2b9eb7868b1577dd3c24b19812a689