Update patch set 1
Patch Set 1:

the glance upload never completes. Try for example in localrc:

IMAGE_URLS="http://launchpad.net/cirros/trunk/0.3.0/+download/cirros-0.3.0-x86_64-disk.img"

You end up with an image stuck in saving:

| 8e998f3f-931e-458f-9446-f617550d1242 | cirros-0.3.0-x86_64-disk | SAVING |        |

works fine with the fix here.

The man page for zcat --force says:

If the input data is not in a format recognized by gzip, and if the option --stdout is also given, copy the input data without change to the standard output: let zcat behave as cat.

so perhaps it doesn't fallback because we aren't using --stdout ?

Patch-set: 1