Update patch set 8
Patch Set 8: (1 inline comment)

I had just written up my in-line comment basically asking the same question - should the user setup the interface on the bridge (and devstack not delete the bridge) or should devstack take care of all this. I think the former is simpler (for devstack at least). Maybe that's OK for now, especially of the single-node case doesn't require attaching a physical interface to the bridge.

Patch-set: 8