Update patch set 8
Patch Set 8:

I agree with Dan here about adding/removing the physical interface. It's more than just moving the ip address from one interface to the other. (Assuming the host is using dhcp we would need to kill dhclient running on eth0 and then start it on the bridge interface. I also believe that NetworkManager will restart the dhclient process again on eth0). I vote for figuring this out and adding it to lib/quantum later if this is needed.

Patch-set: 8