Update patch set 9
Patch Set 9:

I just saw the part about configuring veths to bridge between br-ex and the external network in the commit message. Can you explain why that is needed in typical devstack scenarios, as opposed to simply putting the public IPs on the same bridge that the physical interface is on? I'm sure I'm missing something basic here, but want to understand why it needs to be done this way.

Patch-set: 9