improved Quantum plugin configuration
The configuration defaults for the openvswitch plugin are changing in
https://review.openstack.org/#/c/12362/ to address
https://bugs.launchpad.net/quantum/+bug/1045142. To summarize, with no
overriding of default configuration values, tenant networks will now
work on all systems, but are now local to the host. Using GRE tunnels
or VLANs for external connectivity requires additional
configuration. This patch provides and documents a set of simple shell
variables that can be set in localrc to achieve a range of quantum
network configurations.

Note that this patch is WIP, since the comments are updated for
linuxbridge, but only openvswitch is fully implemented.

Change-Id: I72e9aba1335c55077f4a34495e2d2d9ec1857cd5