Update patch set 1
Patch Set 1:

Just to make it clear: I volunteer to deal with openSUSE/SLE issues in devstack; just redirect people and assign bugs to me.

Also, as mentioned in the commit message: this first patch is missing package names, so the openSUSE support cannot really be used at the moment. This is because I have a branch that implements package maps (ie, we only define the list of packages we need for debian, and then we have mappings from debian names to fedora and opensuse names). I intend to submit that for review afterwards. Alternatively, we can add a simple package list for openSUSE like we do right now for other distros, but I don't think it scales too well...

Patch-set: 1
Label: Code-Review=0