Configure and launch Heat REST API.
This is a REST API in the same style as other OpenStack APIs.

This also creates a new endpoint for the REST API which uses the serivce type
'orchestration'. The old endpoint now has the service type 'cloudformation'.

This matches the pattern where the native openstack API gets a generic service
type while the emulated EC2 API gets a specific type (eg, object-store, s3).

There will be breakage for the time period where only one of this change and
https://review.openstack.org/#/c/14263/ are approved, since keystone will have
the incorrect service type for that period.

Change-Id: I6a0d51a63da8017d375b4c065c4c9079dfca8fe3