Download cirros from launchpad and cache it
Fixes bug 1068110.

images.ansolabs.com/cirros.. was not available, so this change modifies
the download location to launchpad.
The image might already been downloaded, and cached under $FILES/images
With this modification, the script is using that cache location.  Also
set the FILES variable at the top of the script

Change-Id: Idbfebce4e797256cf87a8d5ad0f42c0b3efc7148