change Quantum network plugin set part
Currently in devstack, we can set Q_PLUGIN value to one of well-known quantum plugins ("openvswitch", "linuxbridge", "ryu").

I would like to run devstack with Q_PLUGIN set to a custom Quantum plugin, but stack.sh exits with an error if I do so.
I feel that this is quite limiting and I would like to offer a patch that allows custom Quantum plugins to be used with devstack.

Please let me know if someone disagrees.

Change-Id: Iafe07cf162b5b07bee120b0c571149c374004bb3
Signed-off-by: Takaaki Suzuki <suzuki@midokura.com>