refactor out mysql configuration
on the road to making devstack support multiple databases extract
mysql's configuration like openstack core components are extracted
to make it much easier to replicate this with another database like
postgresql.

This does not address all the places where quantum directly touches
mysql because that work is still in flux.

Part of blueprint:postgresql-support

Change-Id: Ie6bea95abefd0836fd020ec0b78fe36b8628d69a