Update patch set 3
Patch Set 3: (3 inline comments)

Hi,
I have added in a few comments. I'll test this tomorrow in more detail.
I am not sure if it is in the scope of these tests but it would be great if we can have 2 sets of tests - with and without namespaces.
This can help us isolate a number of issues.
Thanks for doing this. 
Gary

Patch-set: 3
Label: Code-Review=0
Label: Workflow=0