Update patch set 1
Patch Set 1: (3 inline comments)

I haven't had a chance to try this yet but it is a nice first cut.

Does there need to be a files/rpms/mysql?

Did you remove python-mysqldb because it is picked up in pip-requires?  If so does it need to be in files/apts/mysql?

Patch-set: 1
Label: Code-Review=0