Update patch set 4
Patch Set 4: I would prefer that you didn't merge this

So this doesn't work on 12.04 out of the box because it's missing python-psycopg2 in the package lists.

One thing I was looking at in the other approach was file/apt and file/rpm lists for the databases like for the core packages, so we could move the database client dependencies out of the projects and into the database files. That would prevent having installed python-mysqldb on postgresql installs, and visa versa.

Patch-set: 4
Label: Code-Review=-1
Label: Workflow=0