Update patch set 5
Patch Set 5: Looks good to me (core reviewer)

I'm good with this landing at this point. This is a nice piece of work, and definitely useful.

As follow up I'd like to move the mysql packages (and postgresql-server) down into files as well. But that's pretty minor cleanup past this.

Patch-set: 5
Label: Code-Review=+2
Label: Workflow=0