Update patch set 5
Patch Set 5:

In order for each exercise script to be self-contained you would need to duplicate the setup/teardown in each script when they are running stand-alone.  If you also detect if it is running under exercise.sh skip the setup/teardown and do it in exercise.sh like you have now.

Patch-set: 5