Update patch set 7
Patch Set 7:

Thanks Nachi.

It is only the ping check functions that I have a problem with now and that is a design choice.  Those functions are not generally useful if there is a case where they exit the calling script and that appears to be why you added the $4 argument.  If they just return a status (say 0 for succ and 1 for failure) then the caller can decide how to handle the issue and $4 is not necessary.

Patch-set: 7
Reviewer: Gerrit User 970 <970@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=0
Label: Workflow=0