Update patch set 1
Patch Set 1:

I quickly took a look at rpms/quantum and learned that it is quite different from apts, which is quite reasonable; packaging is different, probably default installers are different.  (I hit this issue with Quantal server, BTW).

So, my vote is that just keep this patch for apt only, modifying the commit log to say so because i'm not sure if this issue happens on rpm based platforms and I don't even know which package arping belongs to.

What do you think?

Patch-set: 1