Update patch set 1
Patch Set 1: Looks good to me (core reviewer)

[This is mostly to remind myself later how to unbreak the image build scripts.]

Usually when services have #NOPRIME it is to record the dependency for building images but we don't want the service started.  This dates back to the image build scripts in tools/*.  I'm not ready to dump those scripts yet although some of them are way out of date.

I created https://bugs.launchpad.net/devstack/+bug/1078870 to track that.

Patch-set: 1
Label: Code-Review=+2