Update patch set 1
Patch Set 1:

It shouldn't be two hard to land because of how the patches are broken up.  The first nova patch adds the service, but nothing uses it.  After that goes in, this patch can go in.  Once this patch is in, we can proceed with the patches that use the new service.

Patch-set: 1
Label: Code-Review=0