Update patch set 1
Patch Set 1:

Maybe i'm missing something, but it seems like that template doesn't have a VM plug directly into the external network, but rather, just creates a floating IP on the external network, which should already be possible with the existing setup.  

can you point out in the template where it creates a vm directly on the external network?  thx.

Patch-set: 1